Rename legacy hooks to hooks3
tox 4 has a different API that implies rewriting almost the all
code base from scratch. The legacy code will be in hooks3 while
the tox 4 code will be in hooks4.
